@@ -3,7 +3,7 @@ package com.quanda.moviedb.utils
 import android.content.Context
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
-import android.support.v7.app.AlertDialog
+import androidx.appcompat.app.AlertDialog
 import com.quanda.moviedb.R
 
 object DialogUtils {
