@@ -1,12 +1,9 @@
 package com.quanda.moviedb.data.local.db
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.Room
-import android.arch.persistence.room.RoomDatabase
-import android.content.Context
+import androidx.room.Database
+import androidx.room.RoomDatabase
 import com.quanda.moviedb.data.local.dao.MovieDao
 import com.quanda.moviedb.data.model.Movie
-import javax.inject.Singleton
 
 @Database(entities = arrayOf(Movie::class), version = 1, exportSchema = false)
 abstract class AppDatabase : RoomDatabase() {
