@@ -1,7 +1,7 @@
 package com.quanda.moviedb.ui.base
 
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.quanda.moviedb.data.remote.BaseException
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
