@@ -1,6 +1,6 @@
 package com.quanda.moviedb.ui.screen.login
 
-import android.arch.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProviders
 import android.os.Bundle
 import android.util.Log
 import com.quanda.moviedb.BR
