@@ -1,113 +1,31 @@
 package com.quanda.moviedb.utils
 
-import android.animation.ValueAnimator
-import android.app.Activity
-import android.app.ActivityManager
-import android.app.AlarmManager
-import android.app.KeyguardManager
-import android.app.Notification
-import android.app.NotificationManager
-import android.app.Service
-import android.app.admin.DevicePolicyManager
-import android.app.job.JobScheduler
-import android.appwidget.AppWidgetManager
-import android.arch.lifecycle.LifecycleOwner
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProviders
-import android.bluetooth.BluetoothManager
-import android.content.ActivityNotFoundException
-import android.content.ClipboardManager
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProviders
 import android.content.Context
-import android.content.Context.ALARM_SERVICE
-import android.content.Context.CLIPBOARD_SERVICE
-import android.content.Context.CONNECTIVITY_SERVICE
-import android.content.Context.DEVICE_POLICY_SERVICE
-import android.content.Context.INPUT_METHOD_SERVICE
-import android.content.Context.JOB_SCHEDULER_SERVICE
-import android.content.Context.KEYGUARD_SERVICE
-import android.content.Context.NOTIFICATION_SERVICE
-import android.content.Context.TELEPHONY_SERVICE
-import android.content.Intent
-import android.content.Intent.ACTION_CALL
-import android.content.Intent.ACTION_SENDTO
-import android.content.Intent.ACTION_VIEW
-import android.content.Intent.EXTRA_EMAIL
-import android.content.Intent.EXTRA_SUBJECT
-import android.content.Intent.EXTRA_TEXT
-import android.content.Intent.FLAG_ACTIVITY_NEW_TASK
-import android.content.Intent.createChooser
-import android.content.SharedPreferences
-import android.content.res.ColorStateList
-import android.databinding.DataBindingUtil
-import android.databinding.ViewDataBinding
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.Matrix
-import android.graphics.Paint
-import android.graphics.Typeface
-import android.graphics.drawable.Drawable
-import android.media.AudioManager
-import android.net.ConnectivityManager
-import android.net.Uri
-import android.os.BatteryManager
-import android.os.Build
-import android.os.Handler
-import android.os.Looper
-import android.support.annotation.BoolRes
-import android.support.annotation.ColorRes
-import android.support.annotation.DrawableRes
-import android.support.annotation.IdRes
-import android.support.annotation.IntegerRes
-import android.support.annotation.LayoutRes
-import android.support.annotation.RequiresApi
-import android.support.annotation.StringRes
-import android.support.design.widget.BaseTransientBottomBar
-import android.support.design.widget.Snackbar
-import android.support.transition.Transition
-import android.support.v4.app.ActivityOptionsCompat
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentTransaction
-import android.support.v4.app.NotificationCompat
-import android.support.v4.content.ContextCompat
-import android.support.v4.graphics.drawable.RoundedBitmapDrawableFactory
-import android.support.v7.app.AppCompatActivity
-import android.telephony.TelephonyManager
-import android.text.Editable
-import android.text.Spannable
-import android.text.SpannableString
-import android.text.SpannableStringBuilder
-import android.text.Spanned
-import android.text.TextPaint
-import android.text.style.ClickableSpan
-import android.text.style.ForegroundColorSpan
-import android.text.style.StyleSpan
-import android.util.DisplayMetrics
-import android.util.Log
-import android.view.LayoutInflater
-import android.view.MenuItem
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.annotation.BoolRes
+import androidx.annotation.ColorRes
+import androidx.annotation.DrawableRes
+import androidx.annotation.IdRes
+import androidx.annotation.IntegerRes
+import androidx.annotation.LayoutRes
+import androidx.annotation.RequiresApi
+import androidx.annotation.StringRes
+import androidx.transition.Transition
+import androidx.core.app.ActivityOptionsCompat
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentTransaction
+import androidx.core.app.NotificationCompat
+import androidx.core.content.ContextCompat
+import androidx.core.graphics.drawable.RoundedBitmapDrawableFactory
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
-import android.view.ViewGroup
-import android.view.ViewTreeObserver
-import android.view.animation.Interpolator
 import android.view.inputmethod.InputMethodManager
-import android.widget.EditText
-import android.widget.ImageView
-import android.widget.TextView
-import android.widget.Toast
-import com.bumptech.glide.Glide
-import com.bumptech.glide.request.target.SimpleTarget
-import java.io.ByteArrayOutputStream
-import java.io.File
-import java.io.FileOutputStream
-import java.io.IOException
-import java.security.MessageDigest
-import java.security.NoSuchAlgorithmException
-import java.text.ParseException
-import java.text.SimpleDateFormat
-import java.util.*
-import java.util.concurrent.TimeUnit
 
 // # Kotlin Extensions
 
